refactor: migrate to poetry
Organize project into crawler, repro, and scripts packages
