+++ LeanRepoScripts.md
@@ -2,14 +2,24 @@
 
 ## Scripts
 
+Scripts can be run from poetry's virual environment by running
+`poetry run <script name> <options>`. 
+
+For example, `poetry run find_sorries --repository austinletson/sorryClientTestRepoMath --cutoff 7`
+
+Scripts can also be run by activating poetry's virtual environment `eval $(poetry env activate)` and running the script directly.
+
 All scripts require a `GITHUB_TOKEN` environment variable
+
+### List of scripts
 
 1. `get_mathlib_contributors.py`: Gets all contributors to mathlib4 and saves
    results to `all_contributors.txt`
 2. `get_lean_repos.py` takes `all_contributors.txt` as input and checks each contributor's repositories for `lakefile.lean`. Ouputs a list of Lean4 repositories to `lean4_repos.txt`
-3. `find_new_sorries.py --cutoff 7` cycles through all repos in `lean4_repos.txt`
+3. <IS THIS DEPRECATED?>`find_new_sorries.py --cutoff 7` cycles through all repos in `lean4_repos.txt`
 and looks for `sorry` statements whose blame date is less than 7 days old. Output to
 `new_sorries.json`.
+4. `find_sorries --repository <repo> --cuttoff <# of days>` pulls sorries for a specific repo.
 
 ## Known issues
