+++ pyproject.toml
@@ -1,0 +1,31 @@
+[project]
+name = "sorryscraper"
+version = "0.1.0"
+description = "Build a continously updating database of sorry statments in public Lean4 repositories"
+authors = [{ name = "Your Name", email = "you@example.com" }]
+license = { text = "Apache License" }
+readme = "README.md"
+requires-python = ">=3.13"
+
+
+[tool.poetry.dependencies]
+gitpython = "*"
+
+[tool.poetry]
+packages = [
+	{ include = "sorryscraper/scripts", from = "src" },
+	{ include = "sorryscraper/crawler", from = "src" },
+	{ include = "sorryscraper/repro", from = "src" },
+]
+
+
+[tool.poetry.scripts]
+find_sorries = "sorryscraper.scripts.find_sorries:main"
+get_lean_repos = "sorryscraper.scripts.get_lean_repos:main"
+get_mathlib_contributors = "sorryscraper.scripts.get_mathlib_contributors:main"
+offline_sorries = "sorryscraper.scripts.offline_sorries:main"
+
+
+[build-system]
+requires = ["poetry-core>=2.0.0,<3.0.0"]
+build-backend = "poetry.core.masonry.api"
