+++ src/sorryscraper/crawler/__init__.py
@@ -1,0 +1,3 @@
+"""
+crawler pulls sorries from Lean4 GitHub repos.
+"""
