+++ src/sorryscraper/repro/__init__.py
@@ -1,0 +1,5 @@
+"""
+repro(duce) builds Lean4 projects locally and leverages Lean tools to reproduce the proof state at a sorry.
+
+Currently repo only supports REPL.
+"""
